refactor: component Instances for Phone Numbers, Addresses and Emails
Component Instances for Phone Numbers, Addresses and Emails

re #1920
