+++ src/patients/GeneralInformation.tsx
@@ -1,38 +1,26 @@
-import { Button, Panel, Checkbox, Alert } from '@hospitalrun/components'
+import { Panel, Checkbox, Alert } from '@hospitalrun/components'
 import { startOfDay, subYears, differenceInYears } from 'date-fns'
-import { produce } from 'immer'
-import React, { useState } from 'react'
+import React from 'react'
 import { useTranslation } from 'react-i18next'
-import { useDispatch } from 'react-redux'
-import { generate } from 'shortid'
-import validator from 'validator'
 
 import DatePickerWithLabelFormGroup from '../components/input/DatePickerWithLabelFormGroup'
 import SelectWithLabelFormGroup from '../components/input/SelectWithLableFormGroup'
-import TextFieldWithLabelFormGroup from '../components/input/TextFieldWithLabelFormGroup'
 import TextInputWithLabelFormGroup from '../components/input/TextInputWithLabelFormGroup'
-import Address from '../model/Address'
-import Email from '../model/Email'
 import Patient from '../model/Patient'
-import PhoneNumber from '../model/PhoneNumber'
-import { addEmptyPhoneNumber, addEmptyEmail, addEmptyAddress } from './patient-slice'
+import Addresses from './contact-info/Addresses'
+import Emails from './contact-info/Emails'
+import PhoneNumbers from './contact-info/PhoneNumbers'
 
 interface Props {
   patient: Patient
   isEditable?: boolean
   onFieldChange?: (key: string, value: string | boolean) => void
+  addEmptyEntryToPatientArrayField?: (key: string) => void
   onObjectArrayChange?: (
     key: number,
     value: string,
     arrayObject: string | boolean,
     type: string | boolean,
-  ) => void
-  onTempObjectArrayChange?: (
-    key: number,
-    value: string,
-    arrayObject: string | boolean,
-    type: string | boolean,
-    objects: any[],
   ) => void
   error?: any
 }
@@ -44,34 +32,9 @@ const GeneralInformation = (props: Props) => {
     isEditable,
     onFieldChange,
     onObjectArrayChange,
-    onTempObjectArrayChange,
+    addEmptyEntryToPatientArrayField,
     error,
   } = props
-  const dispatch = useDispatch()
-
-  const [phoneNumbers, setPhoneNumbers] = useState<PhoneNumber[]>([
-    {
-      id: generate(),
-      phoneNumber: '',
-      type: '',
-    },
-  ])
-
-  const [emails, setEmails] = useState<Email[]>([
-    {
-      id: generate(),
-      email: '',
-      type: '',
-    },
-  ])
-
-  const [addresses, setAddresses] = useState<Address[]>([
-    {
-      id: generate(),
-      address: '',
-      type: '',
-    },
-  ])
 
   const onSelectChange = (event: React.ChangeEvent<HTMLSelectElement>, fieldName: string) =>
     onFieldChange && onFieldChange(fieldName, event.target.value)
@@ -82,45 +45,8 @@ const GeneralInformation = (props: Props) => {
   const onInputElementChange = (event: React.ChangeEvent<HTMLInputElement>, fieldName: string) =>
     onFieldChange && onFieldChange(fieldName, event.target.value)
 
-  const onObjectElementChange = (
-    index: number,
-    event: string,
-    arrayObject: string | boolean,
-    type: string | boolean,
-  ) => onObjectArrayChange && onObjectArrayChange(index, event, arrayObject, type)
-
-  const onTempObjectElementChange = (
-    index: number,
-    event: string,
-    arrayObject: string | boolean,
-    type: string | boolean,
-    objects: any[],
-  ) => onTempObjectArrayChange && onTempObjectArrayChange(index, event, arrayObject, type, objects)
-
   const onCheckboxChange = (event: React.ChangeEvent<HTMLInputElement>, fieldName: string) =>
     onFieldChange && onFieldChange(fieldName, event.target.checked)
-
-  const phoneNumbersLookGood = () => {
-    if (patient.phoneNumbers) {
-      for (let i = 0; i < patient.phoneNumbers.length; i += 1) {
-        if (!validator.isMobilePhone(patient.phoneNumbers[i].phoneNumber)) {
-          return false
-        }
-      }
-    }
-    return true
-  }
-
-  const emailsLookGood = () => {
-    if (patient.emails) {
-      for (let i = 0; i < patient.emails.length; i += 1) {
-        if (!validator.isEmail(patient.emails[i].email)) {
-          return false
-        }
-      }
-    }
-    return true
-  }
 
   const onApproximateAgeChange = (event: React.ChangeEvent<HTMLInputElement>) => {
     let approximateAgeNumber
@@ -299,549 +225,33 @@ const GeneralInformation = (props: Props) => {
       </Panel>
       <br />
       <Panel title={t('patient.contactInformation')} color="primary" collapsible>
-        <Panel title={t('patient.phoneNumber.panel')} color="primary" collapsible>
-          <table className="table table-hover">
-            <thead className="thead-light">
-              <tr>
-                <th>{t('patient.phoneNumber.type')}</th>
-                <th>{t('patient.phoneNumber.phoneNumber')}</th>
-              </tr>
-            </thead>
-            <tbody>
-              {patient.id &&
-                patient.phoneNumbers &&
-                patient.phoneNumbers.map((a: PhoneNumber, index: number) => {
-                  const errorBool = error && error.phoneNumbers && error.phoneNumbers[index]
-                  return (
-                    <tr key={a.id}>
-                      <td>
-                        <SelectWithLabelFormGroup
-                          name="permanentPhoneNumberType"
-                          label={t('patient.phoneNumber.type')}
-                          value={patient.phoneNumbers[index].type}
-                          isEditable={isEditable}
-                          options={[
-                            { label: t('patient.phoneNumber.types.home'), value: 'home' },
-                            { label: t('patient.phoneNumber.types.work'), value: 'work' },
-                            { label: t('patient.phoneNumber.types.temporary'), value: 'temporary' },
-                            { label: t('patient.phoneNumber.types.old'), value: 'old' },
-                            { label: t('patient.phoneNumber.types.mobile'), value: 'mobile' },
-                          ]}
-                          onChange={(event: React.ChangeEvent<HTMLSelectElement>) => {
-                            onObjectElementChange(index, event.target.value, 'phoneNumbers', 'type')
-                          }}
-                        />
-                      </td>
-                      <td>
-                        <TextInputWithLabelFormGroup
-                          name="permanentPhoneNumber"
-                          isRequired
-                          label={t('patient.phoneNumber.phoneNumber')}
-                          isEditable={isEditable}
-                          placeholder="PhoneNumber"
-                          value={patient.phoneNumbers[index].phoneNumber}
-                          onChange={(event: React.ChangeEvent<HTMLInputElement>) => {
-                            onObjectElementChange(index, event.target.value, 'phoneNumbers', false)
-                          }}
-                          feedback={t(errorBool)}
-                          isInvalid={!!errorBool}
-                          type="tel"
-                        />
-                      </td>
-                    </tr>
-                  )
-                })}
-              {!patient.id &&
-                phoneNumbers?.map((p: PhoneNumber, index: number) => {
-                  const errorBool = error && error.phoneNumbers && error.phoneNumbers[index]
-                  return (
-                    isEditable && (
-                      <tr key={p.id}>
-                        <td>
-                          <SelectWithLabelFormGroup
-                            name="temporaryPhoneNumberType"
-                            label={t('patient.phoneNumber.type')}
-                            value={p.type}
-                            isEditable={isEditable}
-                            options={[
-                              { label: t('patient.phoneNumber.types.home'), value: 'home' },
-                              { label: t('patient.phoneNumber.types.work'), value: 'work' },
-                              {
-                                label: t('patient.phoneNumber.types.temporary'),
-                                value: 'temporary',
-                              },
-                              { label: t('patient.phoneNumber.types.old'), value: 'old' },
-                              { label: t('patient.phoneNumber.types.mobile'), value: 'mobile' },
-                            ]}
-                            onChange={(event: React.ChangeEvent<HTMLSelectElement>) => {
-                              const type = event.target.value
-                              setPhoneNumbers((currentPhoneNumber) =>
-                                produce(currentPhoneNumber, (v) => {
-                                  v[index].type = type
-                                }),
-                              )
-                              onTempObjectElementChange(
-                                index,
-                                type,
-                                false,
-                                'phoneNumbers',
-                                phoneNumbers,
-                              )
-                            }}
-                          />
-                        </td>
-                        <td>
-                          <TextInputWithLabelFormGroup
-                            name="temporaryPhoneNumber"
-                            isRequired
-                            label={t('patient.phoneNumber.phoneNumber')}
-                            isEditable={isEditable}
-                            placeholder="PhoneNumber"
-                            value={p.phoneNumber}
-                            onChange={(event: React.ChangeEvent<HTMLInputElement>) => {
-                              const phoneNumber = event.target.value
-                              setPhoneNumbers((currentPhoneNumber) =>
-                                produce(currentPhoneNumber, (v) => {
-                                  v[index].phoneNumber = phoneNumber
-                                }),
-                              )
-                              onTempObjectElementChange(
-                                index,
-                                phoneNumber,
-                                'phoneNumbers',
-                                false,
-                                phoneNumbers,
-                              )
-                            }}
-                            feedback={t(errorBool)}
-                            isInvalid={!!errorBool}
-                            type="tel"
-                          />
-                        </td>
-                      </tr>
-                    )
-                  )
-                })}
-            </tbody>
-          </table>
-          <div className="addButtonWrapper">
-            {isEditable &&
-              onTempObjectElementChange &&
-              (phoneNumbersLookGood() ? (
-                <Button
-                  outlined
-                  className="addButton"
-                  color="success"
-                  icon="add"
-                  iconLocation="left"
-                  onClick={() => {
-                    if (patient.phoneNumbers) {
-                      if (patient.id) {
-                        dispatch(
-                          addEmptyPhoneNumber(
-                            patient.id,
-                            {
-                              id: generate(),
-                              phoneNumber: '',
-                              type: '',
-                            } as PhoneNumber,
-                            patient.phoneNumbers,
-                            patient.emails,
-                            patient.addresses,
-                          ),
-                        )
-                      } else {
-                        patient.phoneNumbers = patient.phoneNumbers.concat({
-                          id: generate(),
-                          phoneNumber: '',
-                          type: '',
-                        })
-                      }
-                    } else if (patient.id) {
-                      dispatch(
-                        addEmptyPhoneNumber(
-                          patient.id,
-                          {
-                            id: generate(),
-                            phoneNumber: '',
-                            type: '',
-                          } as PhoneNumber,
-                          patient.phoneNumbers,
-                          patient.emails,
-                          patient.addresses,
-                        ),
-                      )
-                    }
-                    if (!patient.id) {
-                      setPhoneNumbers([
-                        ...phoneNumbers,
-                        {
-                          id: generate(),
-                          phoneNumber: '',
-                          type: '',
-                        },
-                      ])
-                    }
-                  }}
-                >
-                  {t('patient.phoneNumber.addPhoneNumber')}
-                </Button>
-              ) : (
-                <Button color="warning" disabled>
-                  {t('patient.phoneNumber.disabledPhoneNumber')}
-                </Button>
-              ))}
-          </div>
-        </Panel>
+        <PhoneNumbers
+          patient={patient}
+          error={error}
+          isEditable={isEditable}
+          onObjectArrayChange={onObjectArrayChange}
+          addEmptyEntryToPatientArrayField={addEmptyEntryToPatientArrayField}
+        />
 
         <br />
 
-        <Panel title={t('patient.email.panel')} color="primary" collapsible>
-          <table className="table table-hover">
-            <thead className="thead-light">
-              <tr>
-                <th>{t('patient.email.type')}</th>
-                <th>{t('patient.email.email')}</th>
-              </tr>
-            </thead>
-            <tbody>
-              {patient.id &&
-                patient.emails &&
-                patient.emails.map((a: Email, index: number) => {
-                  const errorBool = error && error.emails && error.emails[index]
-                  return (
-                    <tr key={a.id}>
-                      <td>
-                        <SelectWithLabelFormGroup
-                          name="permanentEmailType"
-                          label={t('patient.email.type')}
-                          value={a.type}
-                          isEditable={isEditable}
-                          options={[
-                            { label: t('patient.email.types.home'), value: 'home' },
-                            { label: t('patient.email.types.work'), value: 'work' },
-                            { label: t('patient.email.types.temporary'), value: 'temporary' },
-                            { label: t('patient.email.types.old'), value: 'old' },
-                            { label: t('patient.email.types.mobile'), value: 'mobile' },
-                          ]}
-                          onChange={(event: React.ChangeEvent<HTMLSelectElement>) => {
-                            onObjectElementChange(index, event.target.value, 'emails', 'type')
-                          }}
-                        />
-                      </td>
-                      <td>
-                        <TextInputWithLabelFormGroup
-                          name="permanentEmail"
-                          label={t('patient.email.email')}
-                          isEditable={isEditable}
-                          placeholder="Email"
-                          value={a.email}
-                          onChange={(event: React.ChangeEvent<HTMLInputElement>) => {
-                            onObjectElementChange(index, event.target.value, 'emails', false)
-                          }}
-                          type="email"
-                          feedback={t(errorBool)}
-                          isInvalid={!!errorBool}
-                        />
-                      </td>
-                    </tr>
-                  )
-                })}
-              {!patient.id &&
-                emails?.map((p: Email, index: number) => {
-                  const errorBool = error && error.emails && error.emails[index]
-                  return (
-                    isEditable && (
-                      <tr key={p.id}>
-                        <td>
-                          <SelectWithLabelFormGroup
-                            name="temporaryEmailType"
-                            label={t('patient.email.type')}
-                            value={p.type}
-                            isEditable={isEditable}
-                            options={[
-                              { label: t('patient.email.types.home'), value: 'home' },
-                              { label: t('patient.email.types.work'), value: 'work' },
-                              { label: t('patient.email.types.temporary'), value: 'temporary' },
-                              { label: t('patient.email.types.old'), value: 'old' },
-                              { label: t('patient.email.types.mobile'), value: 'mobile' },
-                            ]}
-                            onChange={(event: React.ChangeEvent<HTMLSelectElement>) => {
-                              const type = event.target.value
-                              setEmails((currentEmail) =>
-                                produce(currentEmail, (v) => {
-                                  v[index].type = type
-                                }),
-                              )
-                              onTempObjectElementChange(index, type, false, 'emails', emails)
-                            }}
-                          />
-                        </td>
-                        <td>
-                          <TextInputWithLabelFormGroup
-                            name="temporaryEmail"
-                            label={t('patient.email.email')}
-                            isEditable={isEditable}
-                            placeholder="Email"
-                            value={p.email}
-                            onChange={(event: React.ChangeEvent<HTMLInputElement>) => {
-                              const email = event.target.value
-                              setEmails((currentEmail) =>
-                                produce(currentEmail, (v) => {
-                                  v[index].email = email
-                                }),
-                              )
-                              onTempObjectElementChange(index, email, 'emails', false, emails)
-                            }}
-                            type="email"
-                            feedback={t(errorBool)}
-                            isInvalid={!!errorBool}
-                          />
-                        </td>
-                      </tr>
-                    )
-                  )
-                })}
-            </tbody>
-          </table>
-          <div className="addButtonWrapper">
-            {isEditable &&
-              onTempObjectElementChange &&
-              (emailsLookGood() ? (
-                <Button
-                  outlined
-                  className="addButton"
-                  color="success"
-                  icon="add"
-                  iconLocation="left"
-                  onClick={() => {
-                    if (patient.emails) {
-                      if (patient.id) {
-                        dispatch(
-                          addEmptyEmail(
-                            patient.id,
-                            {
-                              id: generate(),
-                              email: '',
-                              type: '',
-                            } as Email,
-                            patient.phoneNumbers,
-                            patient.emails,
-                            patient.addresses,
-                          ),
-                        )
-                      } else {
-                        patient.emails = patient.emails.concat({
-                          id: generate(),
-                          email: '',
-                          type: '',
-                        })
-                      }
-                    } else if (patient.id) {
-                      dispatch(
-                        addEmptyEmail(
-                          patient.id,
-                          {
-                            id: generate(),
-                            email: '',
-                            type: '',
-                          } as Email,
-                          patient.phoneNumbers,
-                          patient.emails,
-                          patient.addresses,
-                        ),
-                      )
-                    }
-                    if (!patient.id) {
-                      setEmails([
-                        ...emails,
-                        {
-                          id: generate(),
-                          email: '',
-                          type: '',
-                        },
-                      ])
-                    }
-                  }}
-                >
-                  {t('patient.email.addEmail')}
-                </Button>
-              ) : (
-                <Button color="warning" disabled>
-                  {t('patient.email.disabledEmail')}
-                </Button>
-              ))}
-          </div>
-        </Panel>
+        <Emails
+          patient={patient}
+          error={error}
+          isEditable={isEditable}
+          onObjectArrayChange={onObjectArrayChange}
+          addEmptyEntryToPatientArrayField={addEmptyEntryToPatientArrayField}
+        />
 
         <br />
 
-        <Panel title={t('patient.address.panel')} color="primary" collapsible>
-          <table className="table table-hover">
-            <thead className="thead-light">
-              <tr>
-                <th>{t('patient.address.type')}</th>
-                <th>{t('patient.address.address')}</th>
-              </tr>
-            </thead>
-            <tbody>
-              {patient.id &&
-                patient.addresses &&
-                patient.addresses.map((a: Address, index: number) => (
-                  <tr key={a.id}>
-                    <td>
-                      <SelectWithLabelFormGroup
-                        name="permanentAddressType"
-                        label={t('patient.address.type')}
-                        value={a.type}
-                        isEditable={isEditable}
-                        options={[
-                          { label: t('patient.address.types.home'), value: 'home' },
-                          { label: t('patient.address.types.work'), value: 'work' },
-                          { label: t('patient.address.types.temporary'), value: 'temporary' },
-                          { label: t('patient.address.types.old'), value: 'old' },
-                          { label: t('patient.address.types.billing'), value: 'billing' },
-                        ]}
-                        onChange={(event: React.ChangeEvent<HTMLSelectElement>) => {
-                          onObjectElementChange(index, event.target.value, 'addresses', 'type')
-                        }}
-                      />
-                    </td>
-                    <td>
-                      <TextFieldWithLabelFormGroup
-                        name="permanentAddress"
-                        label={t('patient.address.address')}
-                        isEditable={isEditable}
-                        placeholder="Address"
-                        value={a.address}
-                        onChange={(event) => {
-                          onObjectElementChange(index, event.target.value, 'addresses', false)
-                        }}
-                      />
-                    </td>
-                  </tr>
-                ))}
-              {!patient.id &&
-                addresses?.map(
-                  (p: Address, index: number) =>
-                    isEditable && (
-                      <tr key={p.id}>
-                        <td>
-                          <SelectWithLabelFormGroup
-                            name="temporaryAddressType"
-                            label={t('patient.address.type')}
-                            value={p.type}
-                            isEditable={isEditable}
-                            options={[
-                              { label: t('patient.address.types.home'), value: 'home' },
-                              { label: t('patient.address.types.work'), value: 'work' },
-                              { label: t('patient.address.types.temporary'), value: 'temporary' },
-                              { label: t('patient.address.types.old'), value: 'old' },
-                              { label: t('patient.address.types.billing'), value: 'billing' },
-                            ]}
-                            onChange={(event: React.ChangeEvent<HTMLSelectElement>) => {
-                              const type = event.target.value
-                              setAddresses((currentAddress) =>
-                                produce(currentAddress, (v) => {
-                                  v[index].type = type
-                                }),
-                              )
-                              onTempObjectElementChange(index, type, false, 'addresses', addresses)
-                            }}
-                          />
-                        </td>
-                        <td>
-                          <TextFieldWithLabelFormGroup
-                            name="temporaryAddress"
-                            label={t('patient.address.address')}
-                            isEditable={isEditable}
-                            placeholder="Address"
-                            value={p.address}
-                            onChange={(event: React.ChangeEvent<HTMLTextAreaElement>) => {
-                              const address = event.target.value
-                              setAddresses((currentAddress) =>
-                                produce(currentAddress, (v) => {
-                                  v[index].address = address
-                                }),
-                              )
-                              onTempObjectElementChange(
-                                index,
-                                address,
-                                'addresses',
-                                false,
-                                addresses,
-                              )
-                            }}
-                          />
-                        </td>
-                      </tr>
-                    ),
-                )}
-            </tbody>
-          </table>
-          <div className="addButtonWrapper">
-            {isEditable && onTempObjectElementChange && (
-              <Button
-                outlined
-                className="addButton"
-                color="success"
-                icon="add"
-                iconLocation="left"
-                onClick={() => {
-                  if (patient.addresses) {
-                    if (patient.id) {
-                      dispatch(
-                        addEmptyAddress(
-                          patient.id,
-                          {
-                            id: generate(),
-                            address: '',
-                            type: '',
-                          } as Address,
-                          patient.phoneNumbers,
-                          patient.emails,
-                          patient.addresses,
-                        ),
-                      )
-                    } else {
-                      patient.addresses = patient.addresses.concat({
-                        id: generate(),
-                        address: '',
-                        type: '',
-                      })
-                    }
-                  } else if (patient.id) {
-                    dispatch(
-                      addEmptyAddress(
-                        patient.id,
-                        {
-                          id: generate(),
-                          address: '',
-                          type: '',
-                        } as Address,
-                        patient.phoneNumbers,
-                        patient.emails,
-                        patient.addresses,
-                      ),
-                    )
-                  }
-                  if (!patient.id) {
-                    setAddresses([
-                      ...addresses,
-                      {
-                        id: generate(),
-                        address: '',
-                        type: '',
-                      },
-                    ])
-                  }
-                }}
-              >
-                {t('patient.address.addAddress')}
-              </Button>
-            )}
-          </div>
-        </Panel>
+        <Addresses
+          patient={patient}
+          error={error}
+          isEditable={isEditable}
+          onObjectArrayChange={onObjectArrayChange}
+          addEmptyEntryToPatientArrayField={addEmptyEntryToPatientArrayField}
+        />
       </Panel>
     </div>
   )
